fjernet metoder og variabler som ikke brukes
