+++ src/main/java/inf112/skeleton/app/GameController.java
@@ -1,6 +1,5 @@
 package inf112.skeleton.app;
 
-import com.badlogic.gdx.Screen;
 import inf112.skeleton.app.logic.BoardCards;
 import inf112.skeleton.app.objects.IProgramCard;
 import inf112.skeleton.app.objects.PlayerToken;
@@ -9,63 +8,53 @@ import java.util.ArrayList;
 import java.util.HashMap;
 
 public class GameController {
-
     private ArrayList<ArrayList<IProgramCard>> programCards;
     //player 1 har pos 0 og så videre
     private ArrayList<PlayerToken> playerTokens;
     private int numPlayers;
-    private RoboRally roboRally;
-    private BoardCards boardCards;
     private int turns;
     private HashMap<Integer, ArrayList<IProgramCard>> playersCards;
 
-    public GameController(int numPlayers, RoboRally roboRally){
+    public GameController(int numPlayers){
         this.turns = 0;
         this.numPlayers = numPlayers;
-        this.roboRally = roboRally;
-        boardCards = new BoardCards(roboRally);
-        turn(0);
+        playersCards = new HashMap<>();
     }
 
-    public void startPickingCards(){
-        for (int i = 0; i < numPlayers; i++) {
-            boardCards.newTurn();
-            //donePickingCards(boardCards.);
-        }
-    }
-
-    public void donePickingCards(ArrayList<IProgramCard> cardsCurrentPlayer){
+    public void donePickingCards(ArrayList<IProgramCard> cardsCurrentPlayer, BoardCards boardCards){
         int currentPlayer = turns % numPlayers;
         playersCards.put(currentPlayer, cardsCurrentPlayer);
         turns++;
-        turn(turns % numPlayers);
+        turn(turns % numPlayers, boardCards);
     }
 
-    public void createNewPlayer(){
-        //boardCards.createNewPlayer();
-    }
+    public void turn(int playerTurn, BoardCards boardCards){
+        if (playerTurn == 0){
+            ArrayList<IProgramCard> cards1 = playersCards.get(0);
+            ArrayList<IProgramCard> cards2 = playersCards.get(1);
 
-    public void turn(int playerTurn){
-        Screen screen = getScreen();
-        System.out.println("here");
-        /*
-        while (true) {
-            System.out.println("yolo");
+            System.out.println(cards1.toString());
+            System.out.println(cards2.toString());
+            for (int i = 0; i < 5; i++) {
+                if (cards1.get(0).getPriority() > cards2.get(0).getPriority()) {
+                    doTurn(cards1.remove(0), 0, boardCards);
+                    doTurn(cards2.remove(0), 1, boardCards);
+                }
+                else {
+                    doTurn(cards2.remove(0), 1, boardCards);
+                    doTurn(cards1.remove(0), 0, boardCards);
+                }
+            }
         }
-        */
+    }
+    
+    private void doTurn(IProgramCard programCard, int currentPlayer, BoardCards boardCards) {
+        if (programCard.getDirection() != 0) {
+            boardCards.getBoardLogic().getPlayersList().get(currentPlayer).rotatePlayer(programCard.getDirection());
+        }
+        else
+            for (int i = 0; i < programCard.getMovement(); i++)
+                boardCards.getBoardLogic().getPlayersList().get(currentPlayer).moveInFacingDirection();
     }
 
-    public Screen getScreen(){
-        return boardCards;
-    }
-
-    //private
-    /*
-    trekke programkort
-    robots move
-        *priority
-        * turn
-    lasers
-    checkpoints
-     */
 }
